fix: fix wrong imports in CommentRepository, format spacing, indentation, fix typos and blank lines, rename the methods to more meaningful ones, add jjwt dependencies in the gradle spec

+++ src/main/java/org/trickyplay/trickyplayapi/users/repositories/TPUserRepository.java
@@ -1,10 +1,11 @@
 package org.trickyplay.trickyplayapi.users.repositories;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.trickyplay.trickyplayapi.users.entities.TPUser;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TPUserRepository extends JpaRepository<TPUser, Integer> {
+public interface TPUserRepository extends JpaRepository<TPUser, Long> {
     Optional<TPUser> findByName(String username);
 }
